+++ tests/Model/DTO/EnvelopeTagDataTest.php
@@ -16,7 +16,7 @@ class EnvelopeTagDataTest extends TestCase
             'page' => 1,
             'xPosition' => 10,
             'yPosition' => 100,
-            'recipient' => '/api/envelopes/0211f410-268a-4ac3-ac40-b41ee7647092/recipient/xxx',
+            'recipient' => '/api/envelopes/0211f410-268a-4ac3-ac40-b41ee7647092/recipients/xxx',
             'document' => '/api/envelopes/0211f410-268a-4ac3-ac40-b41ee7647092/documents/xxx',
         ];
 
@@ -38,7 +38,7 @@ class EnvelopeTagDataTest extends TestCase
             'page' => 1,
             'xPosition' => 10,
             'yPosition' => 100,
-            'recipient' => '/api/envelopes/0211f410-268a-4ac3-ac40-b41ee7647092/recipient/xxx',
+            'recipient' => '/api/envelopes/0211f410-268a-4ac3-ac40-b41ee7647092/recipients/xxx',
             'document' => '/api/envelopes/0211f410-268a-4ac3-ac40-b41ee7647092/documents/xxx',
         ];