+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to `gosms` will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## [Unreleased]
+### Added
+- Add new IriTemplate and Iri classes to improve work with IRIs
 
 ## [0.2.1] - 2020-06-16